Updating the commit of ecoevolity.
Updating the commit of ecoevolity that is installed by the
setup_project_env.sh script.

This new version fixes a bug where simphycoeval was crashing when trying
to simulate data under generalized tree models that did not have a fixed
tree.

This update does not affect the simulation results already collected for
analyses where the tree was fixed.