Updating how simphycoeval scripts are managed.
Rather than keeping these scripts in the repo, I'm updating
'create_batch_of_simphycoeval_scripts.jl' and adding
'create_batches_of_simphycoeval_scripts.sh' so that we can reproducibly
regenerate them.

This is nicer, because the simphycoeval scripts have full paths that are
not portable. This is because our cluster does weird things with temp
files/dirs on the compute nodes, which makes it difficult to use
relative directories in bash qsub scripts.