Working on path problem on our HPC cluster.
Using the torque submission system on our Hopper cluster, it is
essentially impossible to use all relative paths within job scripts.
This is because, it is not possible to predict what the working
directory is within a job script, or what PBS_O_WORKDIR will be. Neither
is robust to submitting the script directly versus as part of job array.

Thus, in the job script headers created within 'project_util.jl', we
will check to see if we are running a PBS job, and if so, we will 'cd'
to the absolute path of the job script. Not ideal for portability, but
it works.