Updating commit of ecoevolity installed in setup.
Updating setup_project_env.sh to install a newer commit of ecoevolity.

This new commit of ecoevolity fixes a bug that caused sumphycoeval to
crash under some circumstances. This change only allows the results of
analyses to be summarized by sumphycoeval and does not affect the
simulated datasets or the results of analyses of those datasets in any
way.