Purging all 'locus-100-unlinked-' sims.
Due to a bug in simphycoeval, all the simulations in the output
directories that begin with 'locus-100-unlinked-' are simply duplicates
of the simulations in 'locus-100-linked-'.

These are supposed to be simulations of 100bp loci, where only one
variable character per locus is retained. But, the bug caused
simphycoeval to simulate full data sets.

After updating to the version of ecoevolity that has this fix, these
simulations will need to be re-run.