Updating ecoevolity commit SHA.
Updating the version of ecoevolity that is built and installed for use
in the project. The newer commit fixes a bug in simphycoeval that caused
the simulations of 100bp loci with only one variable site per locus to
be incorrect.

This update to ecoevolity does not affect any other results; simulations
of unlinked characters and 100bp loci where all sites are retained are
not affected by this bug fix.