Starting switch to new branch length distribution.
The non-root internal div times were distributed as beta(1, 1) between 0
and the youngest parent of the nodes mapped to the div time.
This led to some wonky tree shapes, many with many div times near zero.
Changing the config files to a beta(2, 1), which from some prelim
tinkering seems to produce more reasonable trees on average.

This change will require re-running all the simulations and analyses.