Updating script for parsing sim results.
Updating parse_sim_results.jl to parse and write the posterior
probabilities of true shared heights and polytomies.  Because there can
be more than one of each of these in each true tree, these probabilities
are not written to a cell of the TSV files where each row is a sim rep.
Rather, these are written to their own TSV files where each row is a
shared height or polytomy, respectively.