Adding 'cladogram' summaries of gecko results.
I've added an option to 'sumphycoeval' to get the MAP tree as
a 'cladogram' (all of the unique div times are equally spaced).
I'm updating 'setup_project_env.sh' to build the commit of ecoevolity
that has this option.
Updating 'scripts/gekkonid-scripts/sum_runs.sh' to use this option
and output 'cladograms' of the Cyrtodactylus and Gekko results.
Adding the output MAP cladograms in this commit, as well.